refactor: refactor geo-related imports to support package
