+++ pida-core/core-domain/src/main/kotlin/com/pida/flowerspot/FlowerSpot.kt
@@ -1,5 +1,7 @@
 package com.pida.flowerspot
 
+import com.pida.support.geo.GeoJson
+import com.pida.support.geo.Region
 import java.time.LocalDateTime
 
 data class FlowerSpot(
