--- pida-core/core-domain/src/main/kotlin/com/pida/flowerspot/Region.kt
@@ -1,21 +1,0 @@
-package com.pida.flowerspot
-
-enum class Region {
-    SEOUL,
-    GYEONGGI,
-    BUSAN,
-    DAEGU,
-    INCHEON,
-    GWANGJU,
-    DAEJEON,
-    ULSAN,
-    SEJONG,
-    GANGWON,
-    CHUNGBUK,
-    CHUNGNAM,
-    JEONBUK,
-    JEONNAM,
-    GYEONGBUK,
-    GYEONGNAM,
-    JEJU,
-}
